backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
cryptography==44.0.0

# HTTP client
httpx==0.28.1

# Vector DB
qdrant-client==1.13.3

# File parsing
pymupdf==1.25.3
python-docx==1.1.2
pdfplumber==0.11.4

# S3 / MinIO
boto3==1.36.5

# Redis
redis==5.2.1

# Email validation
email-validator==2.2.0

# Multipart uploads
python-multipart==0.0.20

# Dev/Test
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
